+++ src/settings.cpp
@@ -727,7 +727,7 @@ bool MDFNI_SetSettingUI(const char *name, uint64 value)
 {
  char tmpstr[32];
 
- trio_snprintf(tmpstr, 32, "%llu", value);
+ trio_snprintf(tmpstr, 32, "%llu", (unsigned long long)value);
  return(MDFNI_SetSetting(name, tmpstr, FALSE));
 }
